requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.5.0

# AI Partner Chat dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pydantic>=2.0.0
markdown>=3.4.0

# 测试依赖
pytest>=7.0.0
pytest-cov>=4.0.0